git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/branches/leon-tag@6880 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ user/lfs_small.c
@@ -17,6 +17,7 @@ enum { lfs_iterations = 1 };
 #include <stdlib.h>
 #include <errno.h>
 #include <fcntl.h>
+#include <inttypes.h>
 
 #include <sys/stat.h>
 #include <sys/types.h>
@@ -70,7 +71,7 @@ creat_test(int n, int size)
 
     memset(buf, 0xab, size);
 
-    unsigned s = 0 , f = 0 ;
+    uint64_t s = 0 , f = 0 ;
     s = time_msec();
 
     for (i = 0, j = 0; i < n; i ++) {
@@ -122,7 +123,7 @@ creat_test(int n, int size)
 	sync();
 
     f = time_msec();
-    printf("%s: creat took %d msec\n",  prog_name,  f - s);
+    printf("%s: creat took %"PRIu64" msec\n",  prog_name,  f - s);
 }
 
 
@@ -133,7 +134,7 @@ write_test(const char *name, int n, int size)
     int i = 0 ;
     int r;
     int fd;
-    unsigned s = 0 , f = 0 ;
+    uint64_t s = 0 , f = 0 ;
     
     s = time_msec() ;
     
@@ -155,7 +156,7 @@ write_test(const char *name, int n, int size)
     
     f = time_msec() ;
     
-    printf("write_test: write took %d msec\n", f - s);
+    printf("write_test: write took %"PRIu64" msec\n", f - s);
 }
 #endif
 
@@ -176,7 +177,7 @@ read_test(int n, int size)
     int fd;
     int j;
 
-    unsigned s = 0 , f = 0 ;
+    uint64_t s = 0 , f = 0 ;
     s = time_msec();
     for (i = 0, j = 0; i < n; i ++) {
 
@@ -201,7 +202,7 @@ read_test(int n, int size)
     }
 
 	f = time_msec();
-    printf("read_test: read took %d msec\n", f - s);
+    printf("read_test: read took %"PRIu64" msec\n", f - s);
 }
 
 static void 
@@ -211,7 +212,7 @@ delete_test(int n)
     int r;
     int j;
  
-    unsigned s = 0 , f = 0;
+    uint64_t s = 0 , f = 0;
     s = time_msec();
     for (i = 0, j = 0; i < n; i ++) {
 
@@ -254,7 +255,7 @@ delete_test(int n)
 	sync();
 
     f = time_msec();
-    printf("delete_test: unlink took %d msec\n", f - s);
+    printf("delete_test: unlink took %"PRIu64" msec\n", f - s);
 }
